# Discover faster implementation for `findIndexR`.