Currently it is only in the benchmarks, but we should change the actual implementation as well. This commit also adds a benchmark for `findIndexR` that has to iterate the whole vector

```
Benchmark algorithms: RUNNING...
All
  findIndexR:               OK (0.18s)
    134  μs ±  11 μs
  findIndexR_naïve:         OK (0.48s)
    72.6 ms ± 4.4 ms
  findIndexR_manual:        OK (0.23s)
    110  μs ± 5.5 μs
  findIndexR (none):        OK (0.15s)
    4.92 ms ± 416 μs
  findIndexR_naïve (none):  OK (0.54s)
    77.1 ms ± 6.7 ms
  findIndexR_manual (none): OK (0.13s)
    4.05 ms ± 358 μs
```

This is the version from master:
```
Benchmark algorithms: RUNNING...
All                 
  findIndexR:        OK (0.18s)
    136  μs ±  11 μs
  findIndexR_naïve:  OK (0.47s)
    69.8 ms ± 6.5 ms
  findIndexR_manual: OK (0.23s)
    221  μs ±  11 μs
```