@Shimuuar 2023-09-18
Note that implementation differ in rather pathological case: `VG.findIndexR (const False) $ V.replicate 10 (undefined :: Int)`. Current implementation is as strict in element as predicate so expression above will return `Nothing` while string version will return ⊥.

Also benchmark is a bit pathological. It disallows inlining of predicate so GHC has to box value. I wonder how things will change if GHC get to see that predicate is strict and it could pass unboxed double.