@lehins 2023-09-18
> Current implementation is as strict in element as predicate so expression above will return Nothing while string version will return ⊥.

This benchmark is for unboxed vectors which can't store bottom. This means we could use this implementation for Unboxed, Primitive and Storable, but not for Boxed, which is already slower anyways. Am I wrong about this?


None of the combinations of `INLINE` and `INLINEABLE` made a difference on performance, eg this benchmark:
```haskell
, bench "findIndexR_inlined" $ nf findIndexR_inline ( \ !x -> x < indexFindThreshold, as)
```
```haskell
findIndexR_inline :: (Double -> Bool, Vector Double) -> Maybe Int
{-# INLINE findIndexR_inline #-}
findIndexR_inline (pred, v) = go $ V.length v - 1
  where go i | i < 0                              = Nothing
             | inline (pred (V.unsafeIndex v i))  = Just i
             | otherwise                          = go $ i-1
```
has the same performance as the one without any inlining on master.