@Shimuuar 2023-09-19
It doesn't change anything because `nf` prevents inlining anyway. Below are benchmarks result on my machine (commit 5935afafa681da5ba87c3a48219786a7f4147a33):
```
  findIndexR:               OK (0.26s)
    187  μs ±  11 μs
  findIndexR_manual:        OK (0.19s)
    169  μs ±  12 μs
  findIndexR (none):        OK (0.21s)
    6.79 ms ± 419 μs
  findIndexR_manual (none): OK (0.20s)
    6.16 ms ± 429 μs
```

However. If I allow partial application of predicate like:
```haskell
findIndexR :: (Double -> Bool) -> Vector Double -> Maybe Int
{-# INLINE findIndexR #-}
findIndexR_manual :: (Double -> Bool) -> Vector Double -> Maybe Int
{-# INLINE findIndexR_manual #-}
```
```haskell
    , bench "findIndexR" $ nf (findIndexR (<indexFindThreshold)) as
```
Results change drastically:
```
  findIndexR:               OK (0.18s)
    53.4 μs ± 5.3 μs
  findIndexR_manual:        OK (0.23s)
    209  μs ±  11 μs
  findIndexR (none):        OK (0.20s)
    1.53 ms ±  86 μs
  findIndexR_manual (none): OK (0.12s)
    7.65 ms ± 698 μs
```

I'm not sure why hand rolled loop performed so much worse. I didn't look into core. But it seems that performance gains of handrolled loop are highly situational