@lehins 2023-09-19
Alright, than I think it makes sense to abandon this idea of optimizing `findIndexR`, especially since even the gains that I observed were tiny.

@Shimuuar Thank you for investigating it a bit further!